Tests/LibWeb/Text/expected/Editing/execCommand-no-recursion.txt

input event
inner result: false
outer result: true
foo
